Move PatternGenerator::WalkTree into new class TreeWalk.
TreeWalk is a function object - for the sole reason that I need access to
fNlevels, fNlevels can't be a global without breaking threading, and I really
don't want to pass such a constant as a function parameter through all levels.

Also - the parameter of the action() object is now a const NodeDescriptor&,
which contains the parameters to convert a base pattern into an actual pattern.
This ensures thread safety if several threads should operate on the same tree,
plus it's cleaner.