Significant change: To reduce the number of patterns generated by single hits in plane pairs (like u1/u1p), no longer set pattern bits in a reference plane ("layer") located halfway between the two planes. Instead, use the wire planes as pattern planes, regardless of whether or not there is a partner plane.  This may increase the number of planes needed for the pattern recognition (e.g. from 3 to 5 for the 3x5-plane GEn setup). If the plane number increases, the pattern recognition will run slower (time roughly proportional to the plane number increase), but generating the hitpatterns will be faster as fewer bits are set for single hits.
Testing shows that the effect on the number of patterns (and on the
speed) is as expected (about 1/2 the patterns and no longer a tail of
events with very large >100 number of patterns; zero net effect on
overall speed). However, the change also exposes a weakness in the
pattern clustering algorithm in Projection::MakeRoads() - probably the
reason for the large number of "multitracks" people are seeing with
the 3x4-plane configuration.