Properly take the x/y coordinates of the wire plane position into account when calculating hit positions. These coordinates were previously ignored. This allows specifying alignment offsets of individual wire planes via the "position" parameter in the database.
If a wire plane has an x/y offset, it is assumed to be the offset relative to
the MWDC (=detector) frame (so that if the MWDC moves, the wire plane
positions stay the same).  Hit positions are always given relative to the
MWDC frame, not the wire plane frame, hence the plane offset must be added.