Add threading support and implement threads for the Projection::Track() stage in CoarseTrack().  Depending on the data and the hardware, this may gain around 25% speed.
By default, threading is enabled whenever more than one CPU is reported by
TSystem::GetSysInfo(), and as many threads are run (up to the number of
projections = 3) as there are CPUs. However, this may actually cause a
performance hit on CPUs running with hyperthreading, i.e. reporting 2 or 4
logical CPUs for 1 or two real cores.  (The treesearch algorithm does
not seem to lend itself well to the assumptions underlying hyperthreading.)
In this case, the number of threads can be limited with the database
key "BB.mwdc.maxthreads = 1" or "= 2", respectively.