Maintain binary compatibility between production code and TESTCODE.
We want to do this to avoid mystery crashes when, say, compiling a
replay script against a TESTCODE'ed library, which has happened in
the past.

This means that member variables and virtual functions only needed
for TESTCODE must stay in production code. Processing of these
variables/functions remains disabled in production code, so there is
no CPU overhead. However, production code now needs a negligible
amount of additional memory.