More improvements to thread support.
Change the maxthread logic back to giving the value from the database
priority, and to use the number of CPUs/cores as a fallback. Improve
the messages the users sees, and get rid of the arbitrary "sanity limit"
of the old code. Instead, cap the value at the maximum supported by
ThreadCtrl. (This was a bug in previous versions. The thread code would
misbehave on machines with 32 or more cores - oops.)

The default is to use the number of CPUs/cores, so users normally don't
have to do anything to get multithreading. However, if you have 3 projections
but only 2 cores, you might want to set maxthreads = 1 since it turns out
to be more efficient in this case. Then again, 2 core machines are getting
rare ...

Also, fall back single-threaded processing if the thread library cannot
be loaded.