Make WirePlane a derived class of Plane. Complete but not yet tested.
DefineVariables looks rather messy because of the mostly redundant
definitions of the hit variables (regular vs MC) - we need to check if we
can use common definitions for the WireHit parts.

MC hits do not yet have any actual MC data in them.

Some of the "coord.*" variables are commented out. To use these, we will
have to create a class hierarchy for FitCoord as well, parallel to the hits.