Add support for rotation of Tracker coordinate system wrt global (lab) frame.
The rotation parameters must be set in one of the Init routines, typically
in ReadDatabase or ReadGeometry. Nothing needs to be done if there is no
rotation, i.e. the change is backwards-compatible.

fOrigin of the tracker is defined in the global (unrotated) reference frame:
v_tracker = fRotation.Inverse() * (v_global - fOrigin)
v_gobal = fRotation * v_tracker + fOrigin

For efficiency, fIsRotated is set at the end of Tracker::Init() so that
we don't have to do the relatively expensive fRotation.IsIdentity() check
for every new track.