Change the definition of SoLID::GEMPlane::fOrigin to match libsolgem.
libsolgem defines the origin as the center of the wedge's bounding box.
If the wedge is offset by phioff, the origin needs to be rotated
correspondingly. libsolgem hits are always in the wedge frame (or rather
the strip frame, which shares its origin with the wedge frame), but
we need the hits in the Tracker frame which is shifted by the plane's
origin, including the phioff rotation.