Make projecting origin of final tracks to global z = 0 optional.
By default, tracks are projected to z = 0. This can be turned off via
database flag "proj_to_z0". Not projecting is useful if the front of the
tracker system is positioned at large z, so the projection plane would
be distant.

Change dbconvert to generate a SoLID database with proj_to_z0 = 0.
As the geometry is currently defined for SoLID, the trackers are at
large z, and so the tracks would be projected to the distant vertex,
which isn't very meaningful without accounting for the track curvature
in the field.

Clearly, we need to add proper vertex reconstruction to the SoLID
spectrometer class.