Add logic for dummy planes to Projection
This is the bulk of the new logic. All planes, including dummies, are
in a new vector fAllPlanes with corresponding allowed plane pattern
fAltPlaneCombos. The index of each plane in this vector is stored in
the plane as fAltPlaneNum. fAltPlaneCombos and fAltPlaneNum are used
for TreeSearch only. Track fitting in Roads etc. continues to use
fPlaneNum and fPlaneCombos, as before. fMaxMiss always refers to
non-dummy planes; dummy planes, if defined, are always required
(otherwise they would be more or less pointless).

This is still work in progress. The Plane class decoders should behave
differently in dummy mode.