Consistently set 2-sigma position ranges in Hitpatterns
We define the hit resolution as 1 sigma (as needed for the various
chi2 calculations), so we should cover more than 1 sigma when setting
the hitpattern position range. One might even consider 3 sigma, but
the argument is actually mostly academic since the hitpatterns are
usually binned much coarser than the hit resolution (say, 1mm
vs. 0.1mm) so that factors of 2 or 3 in the resolution only matter at
bin boundaries, and even if a neighboring bin is missed, the one
actually set frequently belongs to a valid pattern as well. Thus, 2 vs
3 sigma here most certainly is not equivalent to going from 95% to
99.8% hit detection efficiency.