When building roads, be more accepting of patterns with missing hits
If the pattern corresponds a missing hit in its front or back plane,
do not check if its respective front or back bin number is "in range".
Edge bin number of patterns with missing hits cover a wider and
harder-to-predict numerical range than patterns constrained by a hit
in the edge planes. This effect is especially pronounced if the front
or back plane is spaced far away from other planes along z.